code was refactored
