+++ src/App.js
@@ -1,18 +1,15 @@
 import React from 'react';
 import { Provider } from 'react-redux';
-import store from './store/createStore';
-import Header from './component/Header/Header';
-import Footer from './component/Footer/Footer';
-import './App.scss'
-import RouteNavigation from './component/RouteNavigation/RouteNavigation';
+import store from './store/store';
+import Header from './components/Header/Header';
+import Main from "./components/Main/Main";
+import Footer from './components/Footer/Footer';
 
 export const App = () => (
   <Provider store={store}>
     <div className='app container-fluid d-flex flex-column'>
       <Header />
-      <main className='row'>
-        <RouteNavigation />
-      </main>
+      <Main />
       <Footer />
     </div>
   </Provider>
