--- src/App.scss
@@ -1,57 +1,0 @@
-@import '~bootstrap/scss/bootstrap';
-@import 'src/assets/scss/func';
-@import 'src/assets/scss/extend';
-@import 'component/common/Spinner/Spinner';
-@import 'component/common/Modal/Modal';
-@import 'src/component/Header/Header';
-@import 'src/component/Home/Home';
-@import 'src/component/Posts/Posts';
-@import 'src/component/AddPost/AddPost';
-@import 'src/component/EditPost/EditPost';
-@import 'src/component/Footer/Footer';
-
-.app {
-  height: 100%;
-  text-align: center;
-  color: #2c3e50;
-  > * {
-    &:first-child,
-    &:last-child {
-      flex-shrink: 0;
-    }
-    &:nth-child(2) {
-      flex-grow: 1;
-      display: flex;
-      flex-direction: column;
-    }
-
-  }
-
-  button:focus {
-    outline: none;
-  }
-}
-
-ul {
-  list-style-type: none;
-  margin: 0;
-  padding: 0;
-}
-
-li {
-  margin: 0;
-  padding: 0;
-}
-
-a {
-  color: $black;
-  text-decoration: none;
-
-  &:hover {
-    color: $danger;
-  }
-}
-
-h1, h2, h3, h4, h5, h6, p {
-  margin: 0;
-}
