+++ src/assets/scss/abstracts/functions.scss
@@ -1,0 +1,4 @@
+@function rem($pixels) {
+  $convertRem: $pixels / 16;
+  @return #{$convertRem}rem;
+}
