+++ src/assets/scss/components/App.scss
@@ -1,0 +1,21 @@
+.app {
+  height: 100%;
+  text-align: center;
+  color: #2c3e50;
+  > * {
+    &:first-child,
+    &:last-child {
+      flex-shrink: 0;
+    }
+    &:nth-child(2) {
+      flex-grow: 1;
+      display: flex;
+      flex-direction: column;
+    }
+
+  }
+
+  button:focus {
+    outline: none;
+  }
+}
