--- src/assets/scss/func.scss
@@ -1,4 +1,0 @@
-@function rem($size) {
-  $remSize: $size / 16;
-  @return #{$remSize}rem;
-}
