+++ src/assets/scss/main.scss
@@ -1,0 +1,20 @@
+@import 'vendors/bootstrap';
+
+@import 'base/reset';
+@import 'base/typography';
+
+@import 'abstracts/functions';
+@import 'abstracts/extend';
+
+@import 'components/App';
+@import 'components/BackgroundAnimation';
+@import 'components/Modal';
+@import 'components/Spinner';
+
+@import 'pages/Header';
+@import 'pages/Home';
+@import 'pages/Posts';
+@import 'pages/AddPost';
+@import 'pages/EditPost';
+@import 'pages/Footer';
+
