+++ src/assets/scss/pages/EditPost.scss
@@ -1,0 +1,38 @@
+.edit-post {
+  flex-direction: column;
+  align-items: center;
+  flex-grow: 1;
+
+  &__form {
+    justify-content: flex-end;
+    align-items: flex-start;
+    flex-grow: 1;
+    width: 75%;
+  }
+  &__input {
+    padding: rem(25) rem(10);
+    &:focus {
+      border-color: $info;
+      box-shadow: none;
+    }
+  }
+  &__textarea {
+    padding: rem(10);
+    &:focus {
+      border-color: $info;
+      box-shadow: none;
+    }
+  }
+  &__btn {
+    @extend %btn;
+  }
+}
+
+@media (max-width: 576px) {
+  .edit-post {
+
+    &__form {
+      width: 90%;
+    }
+  }
+}
