+++ src/assets/scss/pages/Posts.scss
@@ -1,0 +1,25 @@
+.posts {
+
+  &__list {
+    margin: rem(20) 0;
+  }
+  &__item {
+    padding: rem(10) rem(20);
+    border-radius: 5px;
+    box-shadow: 0 0 rem(12) rem(-3) rgba($gray-800, 0.7);
+    transition: box-shadow 300ms ease;
+
+    &:hover {
+      box-shadow: 0 0 rem(25) rem(-3) rgba($info, 0.8);
+    }
+  }
+  &__title {
+    text-align: left;
+  }
+  &__text {
+    text-align: left;
+  }
+  &__btn {
+    border: none;
+  }
+}
