+++ src/assets/scss/vendors/bootstrap.scss
@@ -1,0 +1,1 @@
+@import '~bootstrap/scss/bootstrap';
