--- src/component/EditPost/EditPost.scss
@@ -1,20 +1,0 @@
-.edit-post {
-
-  &__input {
-    padding: rem(25) rem(10);
-    &:focus {
-      border-color: $info;
-      box-shadow: none;
-    }
-  }
-  &__textarea {
-    padding: rem(10);
-    &:focus {
-      border-color: $info;
-      box-shadow: none;
-    }
-  }
-  &__btn {
-    @extend %btn;
-  }
-}
