--- src/component/Footer/Footer.scss
@@ -1,43 +1,0 @@
-.footer {
-  padding: rem(20) rem(20) rem(15);
-  background-color: $gray-300;
-
-  &__auth p {
-    text-align: left;
-  }
-  &__social a:nth-child(n+2) {
-    margin-left: rem(40);
-  }
-  
-  .footer__icon {
-    font-size: rem(40);
-    transition: all 300ms linear;
-
-    &:hover {
-      color: $info;
-      box-shadow: inset 0 50px 20px $dark;
-    }
-  }
-
-}
-
-@media (max-width: 768px) {
-  .footer {
-    &__auth {
-      font-size: rem(13);
-    }
-  }
-}
-
-@media (max-width: 576px) {
-  .footer {
-
-    &__auth p {
-      text-align: center;
-    }
-    &__social {
-      justify-content: center;
-      margin-top: rem(12);
-    }
-  }
-}
