--- src/component/Posts/Posts.scss
@@ -1,19 +1,0 @@
-.posts {
-
-  &__item {
-    box-shadow: 0 0.3rem 0.2rem -0.2rem rgba(52, 58, 64, 0.5);
-    padding-bottom: 1rem;
-    &:last-child {
-      box-shadow: none;
-    }
-  }
-  &__title {
-    text-align: left;
-  }
-  &__body {
-    text-align: left;
-  }
-  &__btn {
-    border: none;
-  }
-}
