+++ src/components/Main/Main.js
@@ -1,0 +1,10 @@
+import React from 'react';
+import RouteNavigation from '../RouteNavigation/RouteNavigation';
+
+const Main = () => (
+  <main className='row'>
+    <RouteNavigation />
+  </main>
+);
+
+export default Main;
