+++ src/components/common/BackgroundAnimation/BackgroundAnimation.js
@@ -1,0 +1,15 @@
+import React from 'react';
+
+const BackgroundAnimation = () => (
+  <div className='hero'>
+    <div className='cube' />
+    <div className='cube' />
+    <div className='cube' />
+    <div className='cube' />
+    <div className='cube' />
+    <div className='cube' />
+    <div className='cube' />
+  </div>
+);
+
+export default BackgroundAnimation;
