+++ src/index.js
@@ -2,7 +2,7 @@ import React from 'react';
 import ReactDOM from 'react-dom';
 import { HashRouter } from 'react-router-dom';
 import 'bootstrap/dist/js/bootstrap.bundle.min';
-import './index.css';
+import './assets/scss/main.scss'
 import { App } from './App';
 
 ReactDOM.render(
